+++ src/components/auth/LoginButton.tsx
@@ -1,12 +1,21 @@
 import React from 'react'
 import { type Provider } from '@/auth/application/domain/Provider'
-import { handleAuthLoginRequest } from '@/auth/controller/handler/AuthLoginRequestHandler'
+import { executeOAuthLoginRequest } from '@/auth/application/usecase/OAuthLoginRequestUseCase'
 
 interface LoginButtonProps {
   className?: string
   provider: Provider
   logo: React.ReactNode
   title: string
+}
+
+async function handleAuthLoginRequest (provider: string): Promise<void> {
+  const url = await executeOAuthLoginRequest(provider)
+  if (url == null) {
+    console.error('로그인 리다이렉트 URL이 없습니다.')
+    return
+  }
+  window.location.href = url.url
 }
 
 export const LoginButton: React.FC<LoginButtonProps> = ({ className, provider, logo, title }) => (
