+++ src/components/auth/SessionProvider.tsx
@@ -5,7 +5,9 @@ import React, { useEffect } from 'react'
 export const SessionProvider = ({ children }: { children: React.ReactNode }): React.ReactElement | null => {
   const { initializeSession } = useSession()
   useEffect(() => {
-    initializeSession()
+    initializeSession().catch(error => {
+      console.error('세션 초기화 실패:', error)
+    })
   }, [])
   return <>{children}</>
 }
