+++ src/system/adapter/inbound/hook/UiContextProvider.tsx
@@ -4,7 +4,7 @@ import React, { useEffect } from 'react'
 
 export default function UiContextProvider ({ children }: { children: React.ReactNode }): React.ReactElement | null {
   const [mounted, setMounted] = React.useState(false)
-
+  console.log('ui context provider 렌더링 횟수 체')
   useEffect(() => {
     setMounted(true)
   }, [])
